+++ kadai1/ebiiim/pkg/conv/conv_test.go
@@ -32,28 +32,28 @@ func TestImgConv_Convert(t *testing.T) {
 	)
 
 	// normal: jpg -> png
-	ic = conv.ImgConv{SrcPath: f1, TgtPath: f2, Options: nil}
+	ic = conv.ImgConv{SrcPath: f1, SrcExt: conv.ImgExtJPEG, TgtPath: f2, TgtExt: conv.ImgExtPNG, Options: nil}
 	err = ic.Convert()
 	if err != nil {
 		t.Errorf(e1)
 	}
 
 	// normal: png -> bmp
-	ic = conv.ImgConv{SrcPath: f2, TgtPath: f3, Options: nil}
+	ic = conv.ImgConv{SrcPath: f2, SrcExt: conv.ImgExtPNG, TgtPath: f3, TgtExt: conv.ImgExtBMP, Options: nil}
 	err = ic.Convert()
 	if err != nil {
 		t.Errorf(e1)
 	}
 
 	// normal: bmp -> tiff
-	ic = conv.ImgConv{SrcPath: f3, TgtPath: f4, Options: nil}
+	ic = conv.ImgConv{SrcPath: f3, SrcExt: conv.ImgExtBMP, TgtPath: f4, TgtExt: conv.ImgExtTIFF, Options: nil}
 	err = ic.Convert()
 	if err != nil {
 		t.Errorf(e1)
 	}
 
 	// non-normal: jpg -> png
-	ic = conv.ImgConv{SrcPath: n1, TgtPath: n2, Options: nil}
+	ic = conv.ImgConv{SrcPath: n1, SrcExt: conv.ImgExtJPEG, TgtPath: n2, TgtExt: conv.ImgExtPNG, Options: nil}
 	err = ic.Convert()
 	if err == nil {
 		t.Errorf(e1)
